adicion de busqueda por nombre y mejora de los inputs de precio

+++ app/src/main/java/com/example/appcorte3/Orders/data/model/ProductForOrder.kt
@@ -4,5 +4,5 @@ import com.example.appcorte3.core.data.local.Product.entities.ProductEntity
 
 data class ProductForOrder(
     val product: ProductEntity,
-    val quantity: Int
+    val quantity: Float
 )
